testing out libquery
First test with pid and using a program to be tracked. The pid does not
yield what I expect it to do. Whereas looking for 'chromium-browser'
only gives me the starting point it seems that no process is actually
calling it. Or at least it's not tracked. I tried to get the
task_info.pid from the process when chromium-browser is launched. But
this does not correlate to the pid I receive from 'htop'. I might be
looking at something which is not the pid or it's just some garbage that
is left there. More investigation needed.