Testing arrays with identifiers
I am not certain if I should create an array with identifiers or if the
provenance actually provides this possibility. I am thinking that it
might work without having to use an array and saving all elements in
there. I need to find a good way how I could look up information though.